s9 : теперь для каждого вида задач будет создаваться свой класс
