--- src/HttpTaskServer/handle/HandleBase.java
@@ -1,4 +1,0 @@
-package HttpTaskServer.handle;
-
-public class HandleBase {
-}
