--- src/HttpTaskServer/handle/HandleTasks.java
@@ -1,13 +1,0 @@
-package HttpTaskServer.handle;
-
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
-
-import java.io.IOException;
-
-public class HandleTasks extends HandleBase implements HttpHandler {
-    @Override
-    public void handle(HttpExchange exchange) throws IOException {
-
-    }
-}
